+++ AsyncInput/AsyncInputPatches.cs
@@ -1,0 +1,85 @@
+﻿using HarmonyLib;
+
+namespace NoStopMod.AsyncInput
+{
+    public static class AsyncInputPatches
+    {
+
+        [HarmonyPatch(typeof(scrConductor), "Update")]
+        private static class scrConductor_Update_Patch
+        {
+            public static void Postfix(scrConductor __instance, double ___dspTimeSong)
+            {
+                if (__instance.isGameWorld)
+                {
+                    long nowTick = NoStopMod.asyncInputManager.currTick - NoStopMod.asyncInputManager.offsetTick;
+                    double calculated = NoStopMod.asyncInputManager.getSongPosition(__instance, nowTick);
+                    __instance.songposition_minusi = calculated;
+                }
+            }
+        }
+
+        [HarmonyPatch(typeof(scrController), "OnMusicScheduled")]
+        private static class scrController_Rewind_Patch
+        {
+            public static void Postfix(scrController __instance)
+            {
+                NoStopMod.asyncInputManager.jumpToOtherClass = true;
+                __instance.conductor.Start();
+            }
+        }
+
+        [HarmonyPatch(typeof(scrConductor), "Start")]
+        private static class scrConductor_Start_Patch
+        {
+            public static bool Prefix(scrConductor __instance, double ___dspTimeSong)
+            {
+                if (NoStopMod.asyncInputManager.jumpToOtherClass)
+                {
+                    NoStopMod.asyncInputManager.jumpToOtherClass = false;
+                    NoStopMod.asyncInputManager.adjustOffsetTick(__instance, ___dspTimeSong);
+                    return false;
+                }
+                return true;
+            }
+        }
+
+        [HarmonyPatch(typeof(scrConductor), "Rewind")]
+        private static class scrConductor_Rewind_Patch
+        {
+            public static void Postfix(scrConductor __instance, double ___dspTimeSong)
+            {
+                NoStopMod.asyncInputManager.adjustOffsetTick(__instance, ___dspTimeSong);
+            }
+        }
+
+        [HarmonyPatch(typeof(scrConductor), "StartMusicCo")]
+        private static class scrConductor_StartMusicCo_Patch
+        {
+            public static void Prefix(scrConductor __instance, double ___dspTimeSong)
+            {
+                NoStopMod.asyncInputManager.adjustOffsetTick(__instance, ___dspTimeSong);
+            }
+        }
+
+        [HarmonyPatch(typeof(scrConductor), "ScrubMusicToTile")]
+        private static class scrConductor_ScrubMusicToTile_Patch
+        {
+            public static void Postfix(scrConductor __instance, double ___dspTimeSong)
+            {
+                NoStopMod.asyncInputManager.adjustOffsetTick(__instance, ___dspTimeSong);
+            }
+        }
+
+        [HarmonyPatch(typeof(scrConductor), "DesyncFix")]
+        private static class scrConductor_DesyncFix_Patch
+        {
+            public static void Postfix(scrConductor __instance, double ___dspTimeSong)
+            {
+                NoStopMod.asyncInputManager.adjustOffsetTick(__instance, ___dspTimeSong);
+            }
+        }
+
+
+    }
+}
